python kernel libraries are accessible from interpreter
This lets the libraries that are accessible for the Python kernel be
accessible from the terminal. In order to launch the Python interpreter,
just run `python-$NAME`, where NAME is the name of the kernel given in
the kernel configuration.

The name given is different so that multiple Python kernels in the same
environment do not clash due to having the same name.