Haskell kernel libraries are accessible from interpreter
This lets the libraries that are accessible for the Haskell kernel be
accessible from the terminal. In order to launch the Haskell interpreter,
just run ghci-$NAME, where NAME is the name of the kernel given in
the kernel configuration. The same works for ghc-$NAME.

The name given is different so that multiple Haskell kernels in the same
environment do not clash due to having the same name.