make possible to generate folder from lockfiles
This makes possible to generate a Jupyter folder from two files: a
package.json file and a yarn.lock one. These two files are able to
generate a static, read-only Jupyter folder. This is not the best in the
general case, since one cannot enable or disable extensions (this is
done by changing files in the folder, which is in the Nix store), but it
can give us a good default folder to be put in Nixpkgs.

The Nix shell also comes with a tool for generating such files. However,
one still must manually change the paths of extensions, and replace the
local paths by ones in the NPM repository. Since these lockfiles are not
supposed to change much, this is an acceptable effort.