Add option for custom ihaskell path
Added an extra input to the ihaskell kernel derivation which
allows users to optionally specify a custom ihaskell derivation
instead of using the one that resides in the input haskellPackages
set. This allows users of haskell.nix to use the packages they
build using that with jupyterWith + ihaskell. The default ihaskell
is still the same as in the current implementation.